+++ cmptype/ChangeLog.md
@@ -1,11 +1,17 @@
 # Changelog for type-sets
 
+## 0.2.0.0 --- 2019-08-11
+
+- Changed the behavior of `CmpType` on type-level tuples to be lexicographic
+- Made usages of `CmpType` produce a compile-time error when the plugin isn't
+    enabled
+
 ## 0.1.0.1 --- 2019-08-06
 
-Fixed a bug in the cabal metadata
+- Fixed a bug in the cabal metadata
 
 ## 0.1.0.0 --- 2019-08-06
 
-Initial release!
+- Initial release!
 
 ## Unreleased changes