+++ magic-tyfams/ChangeLog.md
@@ -1,15 +1,20 @@
 # Changelog for type-sets
 
+## 0.1.1.0 --- 2019-08-11
+
+- Added `WhenStuck`, which plugins can use to offer error messages when the
+    plugin isn't enabled.
+
 ## 0.1.0.2 --- 2019-08-06
 
-Fixed a bug in the cabal metadata
+- Fixed a bug in the cabal metadata
 
 ## 0.1.0.1 --- 2019-08-05
 
-Update documentation
+- Update documentation
 
 ## 0.1.0.0 --- 2019-08-05
 
-Initial release!
+- Initial release!
 
 ## Unreleased changes